requirements.txt
click
pyfiglet
mysql-connector-python